Move tests into directories
