+++ tests/optim/test_scheduler.py
