+++ tests/scripts/inference/test_convert_composer_to_hf.py
