+++ tests/scripts/train/test_train_inputs.py
