+++ tests/tokenizers/test_tiktoken.py
@@ -10,7 +10,7 @@ import transformers
 from llmfoundry.tokenizers.tiktoken import (TiktokenTokenizerWrapper,
                                             bytes_to_unicode)
 from tests.horrible_strings import HORRIBLE_STRINGS
-from tests.test_hf_conversion_script import check_hf_tokenizer_equivalence
+from tests.scripts.inference.test_convert_composer_to_hf import check_hf_tokenizer_equivalence
 
 if TYPE_CHECKING:
     from tiktoken.core import Encoding
